Refactor sockets on server + Maintain increased speed over array implementation + Improve security + Handle multiple active sockets for same user
Expanded the onlineUsers class to accommodate these new additions. We trade space for ease and speed and should think of another approach if this is an issue.

There was an issue where a new message would be broadcast to all online users, even if they did not have a chat with the sender. To fix this, we now only broadcast the new message to the intended recipients -- namely all the sockets on which they are active.

To update online statuses through a socket disconnect without a logout event, we maintain a quick lookup of which user belongs to which socket so we can broadcast a remove offline user event if there are no more open sockets for that user.